+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # @discourser/design-system
 
+## 0.17.0
+
+### Minor Changes
+
+- [#111](https://github.com/Tasty-Maker-Studio/Discourser-Design-System/pull/111) [`669cbbb`](https://github.com/Tasty-Maker-Studio/Discourser-Design-System/commit/669cbbb6083863313618b6365741c9eaf04d3c41) Thanks [@willTastyMakers](https://github.com/willTastyMakers)! - add a new ScenarioComponent
+
+## 0.16.0
+
+### Minor Changes
+
+- [#108](https://github.com/Tasty-Maker-Studio/Discourser-Design-System/pull/108) [`ac55b3e`](https://github.com/Tasty-Maker-Studio/Discourser-Design-System/commit/ac55b3e9e67d4e6306c7cdb38ef63d3aac93411a) Thanks [@willTastyMakers](https://github.com/willTastyMakers)! - add a new ScenarioComponent
+
 ## 0.15.1
 
 ### Patch Changes
@@ -154,7 +166,7 @@
   ```
 
   **Changes:**
-  - Simple components (Button, Badge, Spinner, IconButton, Input, Textarea, Heading, InputAddon, InputGroup) are now directly usable
+  - Simple components (Button, Badge, Spinner, IconButton, Input, Textarea, Header, InputAddon, InputGroup) are now directly usable
   - Button.Group is accessible via the Button namespace for sub-components
   - Compound components (Card, Dialog, Switch, Accordion, etc.) retain their namespace pattern as expected
 
@@ -351,7 +363,7 @@
   ```
 
   **Changes:**
-  - Simple components (Button, Badge, Spinner, IconButton, Input, Textarea, Heading, InputAddon, InputGroup) are now directly usable
+  - Simple components (Button, Badge, Spinner, IconButton, Input, Textarea, Header, InputAddon, InputGroup) are now directly usable
   - Button.Group is accessible via the Button namespace for sub-components
   - Compound components (Card, Dialog, Switch, Accordion, etc.) retain their namespace pattern as expected
 
@@ -548,7 +560,7 @@
   ```
 
   **Changes:**
-  - Simple components (Button, Badge, Spinner, IconButton, Input, Textarea, Heading, InputAddon, InputGroup) are now directly usable
+  - Simple components (Button, Badge, Spinner, IconButton, Input, Textarea, Header, InputAddon, InputGroup) are now directly usable
   - Button.Group is accessible via the Button namespace for sub-components
   - Compound components (Card, Dialog, Switch, Accordion, etc.) retain their namespace pattern as expected
 
@@ -745,7 +757,7 @@
   ```
 
   **Changes:**
-  - Simple components (Button, Badge, Spinner, IconButton, Input, Textarea, Heading, InputAddon, InputGroup) are now directly usable
+  - Simple components (Button, Badge, Spinner, IconButton, Input, Textarea, Header, InputAddon, InputGroup) are now directly usable
   - Button.Group is accessible via the Button namespace for sub-components
   - Compound components (Card, Dialog, Switch, Accordion, etc.) retain their namespace pattern as expected